# backend/apiserver/openapi-spec.yaml
openapi: 3.0.1
info:
  title: trippify API
  description: trippify API
  version: 1.0.0
servers:
  - url: "http://localhost:80"
    description: "Local development server"

paths:
  /api/user/{username}:
    get:
      operationId: user.get
      summary: Get the user with the given username
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        200:
          description: The requested user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        404:
          $ref: "#/components/responses/ErrorResponse"

  /api/newUser:
    post:
      operationId: newUser.post
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUser"
      responses:
        201:
          description: The user has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /api/trip/{tripId}:
    get:
      operationId: trip.get
      summary: Get the trip with the given id
      parameters:
        - name: tripId
          in: path
          required: true
          description: The id of the trip
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        200:
          description: The requested trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        404:
          $ref: "#/components/responses/ErrorResponse"

  /api/newTrip:
    post:
      operationId: newTrip.post
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTrip"
      responses:
        201:
          description: The trip has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"

  /api/trip/{tripId}/spot/add:
    post:
      operationId: addSpotToTrip.post
      summary: Add a spot to a given trip
      parameters:
        - name: tripId
          in: path
          required: true
          description: The id of the trip
          schema:
            $ref: "#/components/schemas/ObjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewSpot"
      responses:
        201:
          description: The spot has been added to the trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"

  /api/spot/{spotId}:
    get:
      operationId: spot.get
      summary: Get the spot with the given id
      parameters:
        - name: spotId
          in: path
          required: true
          description: The id of the spot
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        200:
          description: The requested spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spot"
        404:
          $ref: "#/components/responses/ErrorResponse"

components:
  schemas:
    NewUser:
      type: object
      required:
        - username
        - name
      properties:
        username:
          $ref: "#/components/schemas/Username"
        name:
          type: string

    User:
      allOf:
        - $ref: "#/components/schemas/NewUser"
      type: object
      required:
        - _id
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        trips:
          type: array
          items:
            $ref: "#/components/schemas/ObjectId"

    Username:
      type: string
      pattern: "^[a-z0-9.]+$"

    NewTrip:
      type: object
      required:
        - owner #TODO: Get the user from the auth
        - title
      properties:
        owner:
          $ref: "#/components/schemas/ObjectId"
        title:
          type: string

    Trip:
      allOf:
        - $ref: "#/components/schemas/NewTrip"
      type: object
      properties:
        start_date:
          type: string
        end_date:
          type: string
          format: date-time
        spots:
          type: array
          items:
            $ref: "#/components/schemas/ObjectId"

    NewSpot:
      type: object
      required:
        - title
        - longitude
        - latitude
        - date_from
        - images
      properties:
        title:
          type: string
        longitude:
          type: number
        latitude:
          type: number
        date_from:
          type: string
          format: date-time
        description:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"

    Spot:
      allOf:
        - $ref: "#/components/schemas/NewSpot"
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        date_to:
          type: string
          format: date-time

    Image:
      type: object
      required:
        - source
      properties:
        source:
          type: string
        timestamp:
          type: integer
          format: int32

    ObjectId:
      description: A MongoDB ObjectId
      type: string
      pattern: "^[a-fA-F0-9]{24}$"

    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          description: A human readable error message
          type: string
        code:
          $ref: "#/components/schemas/ErrorCode"

    ErrorCode:
      type: string
      enum:
        - server_error
    Timestamp:
      type: string
      format: iso8601

  responses:
    ErrorResponse:
      description: Error Response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"